requirements.txt
# Dashboard Streamlit pour données météo
streamlit
pandas
requests
geopy
openmeteo-requests
requests-cache
retry-requests
googletrans
blagues-api
asyncio

# Dépendances pour chatbot IA avec agent météo
langchain
langchain-groq
langchain-core
langchain-community
langgraph
python-dotenv
streamlit-chat